setting up for panel entity values
